requirements.txt
django
environs
django-phonenumber-field[phonenumberslite]
django-templated-mail
djangorestframework
django-ipware
ua-parser
user-agents
django_jalali[drf]
django_jalali
django-jalali-date 
django-extensions
django-cors-headers
